chore(cli): lazy-load action modules in program.ts files
Extract action bodies from program.ts into separate files grouped
by import similarity, so that loading a program does not eagerly
import the world.

playwright-core/src/cli/program.ts:
- browserActions.ts: open, codegen, screenshot, pdf (imports playwright client)
- installActions.ts: install, uninstall, install-deps (imports registry/server)
- driver, trace viewer, cli-client: lazy-imported inline

playwright/src/program.ts:
- testActions.ts: test, clear-cache, dev-server, test-server (imports runner)
- reportActions.ts: show-report, merge-reports (imports reporters)
- MCP server, init-agents: lazy-imported inline
