+++ packages/playwright-core/src/cli/DEPS.list
@@ -9,3 +9,5 @@
 ../server/trace/viewer/
 ../tools/cli-client/program.ts
 ../tools/trace/traceCli.ts
+./browserActions.ts
+./installActions.ts
